Agregando primer test y corrigiendo tipos de datos

+++ src/main/java/project/hex/data/DefaultProductGateway.java
@@ -51,7 +51,7 @@ public class DefaultProductGateway implements ProductGateway {
         return ProductEntity.builder()
                 .id(product.getId())
                 .nombre(product.getNombre())
-                .precio((long) product.getPrecio())
+                .precio(product.getPrecio())
                 .build();
     }
 }
