+++ src/main/java/project/hex/domains/Product.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 public class Product {
     private Long id;
     private String nombre;
-    private double precio;
+    private Long precio;
 }
